Please excuse me Lubos, I reverted your KIPI fix for the following reasons: - It requires SVN versions of KDE and libkipi - It assumes that a KIPI plugin does nothing until it is activated by a   KAction. This might be true for now, but the KIPI design allows the plugin to   do anything as long as it's loaded, so using a place holder action is not a   good thing (*)
I should have been more verbose when you submitted your patch. My fault.

I implemented a much less sophisticated work-around to avoid waiting for KIPI
plugins to load. It uses the QTimer::singleShot trick along with querying the
BusyLevelManager if we are not too busy to load the plugins.

It has the drawback that the menu and accelerators won't work in the few
seconds between the start of the application and the actual loading of the
plugins, but I consider this a very minor issue compared to the speed gain.

* : I actually was against this KIPI design, and suggested associating KIPI
plugins with .desktop files which would have described them. This would have
made it much easier to do on-demand loading.

CCMAIL: gwenview-general@lists.sf.net

svn path=/trunk/extragear/graphics/gwenview/; revision=428829