Fix template instantiation for dll export on Win32
Since MSVC can't export templated classes from a DLL, we need
to make sure that every user of the BaseBinder/Binder classes
instantiates the template by himself. As the class is completely
contained in the .h file, this is not a problem.
I hope this doesn't break anything for gcc (it shouldn't...)