Make "lock zoom" feature optional
Having used Gwenview with the lock zoom feature back in. I realize it
is a useful feature, but strongly believes should not be enabled by default.

This commit reverts to not locking zoom by default, but introduces an
hidden config option which can be set with:

kwriteconfig --file gwenviewrc --group ImageView \
  --key ShowLockZoomButton --type bool true

When this is set, a lock button appears on the right of the zoom level.
Clicking it will lock zoom, clicking it again will unlock zoom.

I want to find a better solution for this in KDE SC 4.10: I don't like how
the lock button adds clutter to the already cluttered, zoom widgetry, but
this should do for KDE SC 4.9.

CCBUG: 291759