Fix weird behavior when dropping an image in the thumbnail view
Before this change, if the thumbnail view had an image selected, attempting
to drop an image in it, would result in a new drag being started using the
selected image.

There is apparently no need to turn mouse tracking on for the thumbnail
speed bar anymore, so just disable this code.