Write thumbnails to disk more frequently
This patch starts ThumbnailCache every time an image is queued.

Up till now ThumbnailCache was started at the destructor of ThumbnailLoadJob. I think this is suboptimal since it causes a lot of waiting in the destructor when all thumbnails are written to disk at once. (This blocks the UI.) Instead of that I trigger ThumbnailCache every time an image is queued.