Bring back thumbnail buttons with Plastique or Cleanlooks style
The Qt::WA_Hover attribute was removed in commit 9c6ce9d5 to fix a
drag'n'drop bug, but setting WA_Hover was not the cause of the bug. The
problem came from the call to setMouseTracking().

Note that I set the attribute in the delegates instead of in the view,
this causes code duplication, but it makes more sense for the delegates to
set the attributes they need than to leak delegate needs in the view.

The code handling the thumbnail buttons should be moved out of the
delegates to avoid duplication, but that can't be done for 4.10.

REVIEW: 108660
FIXED-IN: 4.10.0