Move code checking next image is selected after removal to ContextManager
This is necessary after this fix to KLinkItemSelectionModel [1]. Without
this change Gwenview does not automatically advance to the next image
when current image is removed.

It makes more sense anyway since there are multiple instances of
ThumbnailView using the same model (or a proxy of it). Right now when a
row of the model is removed every ThumbnailView instance checks the
status of the current index. With this change, the check is done by the
ContextManager, which exists only once.

[1]: https://git.reviewboard.kde.org/r/111953/