Apply display color profile to all images
...including those which do not have any embedded profile. Just treat
them as sRGB, which is probably a reasonable choice. This makes a huge
difference on wide-gamut displays.

LittleCMS doesn't support RGBA_Premultiplied because premultipled alpha
doesn't "make a lot of sense with a color management system since color
perception varies with intensity and the colorspace may not be linear
across the color channels" (see [1] for why, and for an experimental
patch which apparently didn't move anywhere and wasn't merged).

[1] http://sourceforge.net/p/lcms/mailman/message/30246542/

REVIEW: 123928
BUG: 322149
BUG: 341943