Fix handling of multiple arguments in command line
Summary:
When running gwenview with multiple local images as arguments in the
command line, the constructed urls used to create the links in the
temporary directory were wrong. This commit fixes the urls so they
can point to local files.

Also, if some file was specified multiple times in the command line,
the user was presented with a KIO rename dialog, which is quite
strange since the user was just opening some images, so this
commit also removed duplicated arguments.

Reviewers: apol

Reviewed By: apol

Subscribers: apol

Differential Revision: https://phabricator.kde.org/D2880