Scroll wheel and touchpad zoom in smaller increments
Summary:
BUG: 307637

Change the zoom speed so that the scroll wheel and touchpad zoom behavior matches that of Inkscape and Krita: for more than 100%, the zoom is increased by sqrt(2); for less than 100%, the zoom is decreased by sqrt(0.5). In practice this means the following zoom levels:

35% → 50% → 71% → 100% → 141% → 200% → 283% → 400% → 566% → 800%

This feels much more comfortable and offers more control compared to the current ever-accelerating zoom speed.

Submitting this patch on behalf of  Vangelis Tasoulas.

Test Plan: Tested in KDE Neon. Works as advertised.

Reviewers: #kde_applications, vtasoulas, lukas, gateau, rkflx

Reviewed By: vtasoulas, rkflx

Subscribers: ltoscano, anthonyfieroni, rkflx

Differential Revision: https://phabricator.kde.org/D7972