General HiDPI enablement for Gwenview's importer
Summary:
This is a follow-up patch to 43f6e15237f3, which enabled HiDPI in the UI
of Gwenview's main app.

Note: Thumbnails are still pixelated as this is not yet fixed in the
respective Gwenview component.

Solves issue 2. in D7581#163655.

Test Plan:
Standard icons all over the UI not pixelated anymore for
`QT_SCALE_FACTOR=1.5 gwenview_importer .`

Reviewers: gladhorn, hetzenecker

Reviewed By: hetzenecker

Subscribers: ngraham

Differential Revision: https://phabricator.kde.org/D8918