Rename setTargetUrl to setTargetDirUrl
This change was initially meant to be part of D8785, but had been backed
out due to concerns about breaking source compatibility. Meanwhile, in
D8785#176817 it was determined that SC/BC most likely won't be an issue
in Gwenview because of no other users of the code.

Therefore we now can rename a couple of methods and variables to better
reflect what they are about, improving readability.