Resize top bar in fullscreen view mode on toggling sidebar
Summary:
When viewing an image in fullscreen mode and pressing {key F4} to
toggle the sidebar, the toolbar at the top does not resize.
The buttons on the right side are not visible while sidebar shows.
In browse mode `KToolBar` is used with a standard layout where this
problem does not appear. Since the toolbar in view mode overlaps the
image view a standard layout is not usable here which would send
resize events automatically.
This patch adds a resize Eventlistener to adjust toolbar width.

BUG: 387784

Test Plan:
1. Open gwenview in fullscreen view mode
2. Toggle sidebar ({key F4})
-> The top toolbar should resize according to image view width

Checked to work properly on a second monitor.

Reviewers: rkflx, ngraham

Reviewed By: rkflx

Differential Revision: https://phabricator.kde.org/D10248