Fix squished toolbar buttons in fullscreen view mode
Summary:
In fullscreen view mode the sidebar splitter can be moved to the right
until the toolbar buttons are squished a bit.
This is caused by a missing layout connection which would propagate
the correct minimum size hints. Since the toolbar in view mode
overlaps the image view a standard layout is not usable here.
This patch sets a minimum size hint for the image view according to
the internal layout of the overlay toolbar width.

{F5690953}

Test Plan:
1. Open gwenview in fullscreen view mode
2. Move sidebar splitter to right just before collapsing
-> The buttons of the top toolbar should remain in original size

Checked to correctly adjust width while:
* Toggling status bar ({key F3})
* Toggling {nav Show thumbnails}
* Opening images with different length filenames (thumbnails off)

Reviewers: rkflx, ngraham, #gwenview

Reviewed By: rkflx

Differential Revision: https://phabricator.kde.org/D10458