Remove undesirable accelerators
Editing `.ui` files with Qt Designer in KDE Plasma can result in
superfluous `&amp;` accelerators being added (see
https://stackoverflow.com/a/32711483 and D10421#206934 for discussion).
This happened in 4a87ed952ec2 and 4efdfcdaef68, with the remaining
accelerators at the start of words as a deliberate choice.

Test Plan:
Even after the removal keyboard access is still possible (ignoring the
bug that for Crop and Red Eye Reduction the underline was not displayed
anyway even though the accelerator worked and still works…).