Save config when opening the config dialog
Summary:
Some settings can be changed outside of the config dialog. We don't
want these changes to be lost when changing the config, which
triggers reloading of the config.

BUG: 390331

Test Plan:
Settings changed outside the config dialog, e.g. thumbnail zoom level in Browse view,
should not reset if you change a setting in the config dialog.

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Differential Revision: https://phabricator.kde.org/D10626