Fix thumbnail hover icon show-in-fullscreen action
Summary:
In non-fullscreen browse mode the thumbnails display four hover icons.
Clicking on the "view in fullscreen" icon switches to fullscreen mode
but not to view mode as expected.
This patch switches to fullscreen view mode.

Test Plan:
Open Gwenview in normal browse mode and hover over thumbnail icons
to see the icons. Click on second icon to view this image
in fullscreen mode.

{F5712905}

Reviewers: #gwenview, ngraham

Reviewed By: #gwenview, ngraham

Subscribers: ngraham, rkflx

Differential Revision: https://phabricator.kde.org/D10651