Adjust sidebar code and config entries to statusbar
Summary:
Setup config entries and default values of sidebar states in
`gwenview.kcfg` and use the generated access functions like it
is done for statusbar in D10577.
From a user's sight the sidebar behaviour and state saving is
not changed.

Test Plan:
Switch browse/view, normal/fullscreen modes.
The sidebar state should always restore to previous setting and
button on statusbar should correspond to this state.
Three different states are (re)stored:
* Browse mode (always equal in normal/fullscreen mode)
* View mode normal (windowed) mode
* View mode fullscreen mode

Switch the different tabs on sidebar and check if the last
activated is restored on next restart.

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: rkflx

Differential Revision: https://phabricator.kde.org/D10687