Improve thumbnail tooltip animation
Summary:
This fixes a cosmetic issue where the tooltip got 'stuck' as long as you
kept moving your mouse around. The trade-off is it more often than not
fades out quicker than it did before, but I believe this isn't a problem.

This prevents the tooltip fade-out animation getting the 500ms delay unless the current animation is not running.
Therefore preventing the animation perpetually gaining a delay if moving the mouse over thumbnails fast enough.

BUG: 245091

## Before:
{F5720935, loop, autoplay}

## After:
{F5720936, loop, autoplay}

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: rkflx

Differential Revision: https://phabricator.kde.org/D10704