Replace manual palette adjustment with background role (fullscreen info label)
Summary:
Instead of customising the palette, then monitoring for changes, we set the background color role of the widget.

This has the benefit of keeping the palette inherited, therefore changes to the system palette still propogate. Not to mention we don't need to watch for palette change events.

Test Plan: The information label in fullscreen (where the filename is displayed) should not significantly change color. A slight change is expected as we are not manually setting it but using a pre-determined role.

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Differential Revision: https://phabricator.kde.org/D10781