Hide fullscreen thumbnailbar when on-canvas image tools are in use
Summary:
In fullscreen mode the thumbnailbar will slide in once the cursor moves
near the top. As this covers parts of the image, it makes the upper
handles of the {nav Crop} tool inaccessible and prevents the user from
applying the {nav Red Eye Reduction} tool on every part of the image.

`setDistractionFreeMode` already provides the necessary wiring which was
supposed to prevent the hot edge from triggering, however it only
disabled the auto-hiding. This is fixed by adding a function and a
variable which disables auto-showing.

Test Plan:
Open image in fullscreen {nav View} mode, press {key Shift C} to trigger
{nav Crop} and move cursor to the top. The thumbnailbar does not
reappear again. Stop the tool, now triggering should resume.

Switching to fullscreen should show the thumbnailbar initially without
the cursor triggering, as before.

Reviewers: #gwenview, huoni

Reviewed By: huoni

Subscribers: ngraham

Differential Revision: https://phabricator.kde.org/D10876