Enlarge small SVGs too
Summary:
For images smaller than the viewport at 100% zoom, Gwenview allows to
scale up the image to {nav Fit} the window if {nav Enlarge smaller images}
is checked in the configuration dialog. So far this only worked for
raster-based images, but not for SVGs.

`mEnlargeSmallerImages` is meant to be handled by the abstract base
class `AbstractImageView`, however when f00cef3c0 moved code over there
from `RasterImageView`, it was missed to also adapt `SvgImageView`. This
can be fixed by implementing the missing function.

In addition, a left-over `bool` is removed.

BUG: 364822
FIXED-IN: 17.12.3

Test Plan:
- Open `icons/hisc-apps-gwenview.svgz`, press {nav Fit}: Small icon shown.
- Check {nav Enlarge smaller images}, press {nav Fit} again: Icon scaled up to fit window.
- Enlarging raster-based images still works fine.

Reviewers: #gwenview, huoni, muhlenpfordt, ngraham

Reviewed By: #gwenview, huoni, muhlenpfordt, ngraham

Subscribers: ngraham, muhlenpfordt, huoni

Differential Revision: https://phabricator.kde.org/D10926