Keep click position in focus on middle-click zoom
Summary:
Clicking on an image in View Mode with the middle mouse button toggles
between `Fit` and `100%` zoom modes. The image zoomed to `100%` is
always scrolled to the image center.
This patch keeps the clicked location in focus as it is done while
zooming with {key Ctrl Scroll}.

FEATURE: 308335

Test Plan:
* Open a big image in View Mode
* Middle-click on different locations over the image
* The `100%` zoom view should keep the clicked location in focus

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: rkflx, huoni

Differential Revision: https://phabricator.kde.org/D11336