MPRIS: for images report the slideshow interval as well as timer resets
Summary:
Setting "mpris:length" "Metadata" property of images matching the interval of the
slideshow, signalling resets of the "Position" and delivering the
current timeout progress in the autoadvance timer, all that allows
MPRIS controllers to show the timely progress in the slideshow.

Supporting the same for video files is left-out for now, that needs
some bigger refactoring to have this information properly modelled
by the slideshow object.

((The Plasma media controller applet of <= 5.12 sadly only shows time
progress information if CanSeek=true. That got fixed for Plasma 5.13,
which will show a read-only progressbar instead of a slider))

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: rkflx

Differential Revision: https://phabricator.kde.org/D11359