Swap position and size in crop tool
Summary:
When one wants to re-use crop settings in gwenview (i.e. position and size), they have to be entered in the following order :
1/ Size
2/ Position

(it is not possible to do it the other way, as position is locked until the size of the crop area is smaller than the size of the image).

However, the settings are displayed in the reverse order, which is not convenient (un-necessary mouse actions).

BUG: 391758

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: muhlenpfordt, rkflx, ngraham

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D11329