Prettify display of URLs in Recent Folders and Places
Summary:
Entries in the {nav Recent Folders} tab on the {nav Start Page}
contained superfluous information like `file://` for local folders and
the user's home directory, taking away valuable space and unnecessarily
eliding the text:
  file:///home/user/Documents/

By abbreviating to `~` and omitting the protocol for local folders, the
entries can look much neater, making them faster to scan for users:
  ~/Documents/

The changes not only affect the {nav Recent Folders} tab, but also what
is displayed in the {nav Recent Files} tab and the hover overlays for
very long entries.

In addition, adding folders to the {nav Places} panel now uses the last
subfolder for naming the entry, bringing the behaviour in line with
Dolphin.

BUG: 383850
FIXED-IN: 18.04.0

Test Plan:
- Inspected {nav Recent Folders} and {nav Recent Files} on the {nav Start Page}.
- In {nav Browse} mode, used {nav File > Add Folder to Places}.
- Tested that remote paths containing `~` are still intact.

Before:
{F5758862}

After:
{F5758864}

Reviewers: #gwenview, muhlenpfordt

Reviewed By: #gwenview, muhlenpfordt

Subscribers: ngraham, muhlenpfordt

Differential Revision: https://phabricator.kde.org/D11493