Align slideshow terminology to MPRIS behaviour
Summary:
9631043c110d introduced MPRIS support to Gwenview. Common MPRIS
controllers often only provide "Pause" in the main interface, with
"Stop" being non-existent or hidden. Therefore "holding" the slideshow
was mapped to "Pause". "Play" either "Resumes" automatically advancing
to the next image, or in non-fullscreen mode switches to fullscreen and
starts playback. Conversely, "Stop" ends playback and exits from
fullscreen.

To keep consistency, we change the tooltip text of the fullscreen button
controlling the playback to "Pause". In addition, the internal `enum` is
adapted accordingly to avoid confusing developers.

Note that the icon on the button still shows a "Pause" symbol as before
the patch.

Ref T8222

Test Plan:
Start Gwenview normally, with `-f` and with `-s`, and observe the
tooltip of the playback button in the top fullscreen toolbar as well as
the button text in non-fullscreen mode (use {nav Configure Toolbars} to
add it) make sense in every situation.

Reviewers: #gwenview, kossebau, muhlenpfordt

Reviewed By: #gwenview, muhlenpfordt

Subscribers: ltoscano, muhlenpfordt, huoni, kde-doc-english

Maniphest Tasks: T8222

Differential Revision: https://phabricator.kde.org/D11495