Fix crop ratio combobox not updating crop when text didn't change
Summary:
For example, "US Letter" is duplicated - one for landscape, one
for portrait. Since the crop was only updated when the combobox
text changed, switching between these two didn't do anything.
This patch adds a connection to the `currentIndexChanged` signal
so that changes like this are detected.

Depends on D11378

Test Plan:
  # Open crop tool > Advanced settings
  # Change Ratio to US Letter (portrait)
  # Change Ratio to US Letter (landscape)
  # Repeat for the ISO ratio

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Differential Revision: https://phabricator.kde.org/D11379