Rename Transparent background config options
Summary:
This makes the meaning of this option more clear. It also uses the
correct term "checkerboard" instead of "check board".

I have left the `RasterImageView::AlphaBackground*` terminology the same
as I believe it makes more sense as is in terms of what the code does.
For example `AlphaBackgroundNone` means we do not draw a background or
draw a transparent background.

Depends on D11630

Before:
{F5778723}

After:
{F5781170}

Test Plan: Ensure all transparent background options continue to function.

Reviewers: #gwenview, rkflx, ngraham

Reviewed By: #gwenview, rkflx

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D11824