Adapt PlaceTreeModelTest autotest to KIO change
Summary:
After e8d799c9b522 needed to account for additional entries in
`KFilePlacesModel`, in KIO's 7e1d2fb84546 the number of default items
returned by the model was changed again, in particular "This Month" and
"Last Month" were removed.

As those entries are not that relevant for Gwenview too, we simply
follow along, meaning apart from changing the number expected in the
test no further changes are necessary.

If in the future more changes are planned in KIO, it might be worth
thinking about removing the test altogether, but for now it serves as a
canary, so any effect on Gwenview can be investigated in a timely
fashion.

Test Plan:
`./tests/auto/placetreemodeltest testListPlaces` does not fail anymore
with KIO 5.45, still works fine with KIO 5.44 (make sure CMake picks up
the change before recompiling).

Reviewers: #gwenview, ngraham

Reviewed By: #gwenview, ngraham

Subscribers: ngraham, huoni

Differential Revision: https://phabricator.kde.org/D12101