Display document count labels in View, Browse, and Full Screen modes
Summary:
{nav Browse} mode: The existing total count is now split into images
and videos, e.g. "5 images, 3 videos". When 2 or more documents (that
is, images or videos) are selected, the label changes to display how
many are selected and their total size. E.g. "4 of 10 selected (3.4
MiB).

{nav View} mode: A new label in the status bar, showing the current
document index, and total in the folder, e.g. "32 of 753". This label is
hidden in Lighttable mode, and elided if there's no room for it.

{nav Full Screen} mode: Similar to {nav View}, the current index and
total count is displayed. This is in the format "32 of 753", and is
anchored to the bottom right of the information panel.

Each mode calculates these values independently. This is because for
{nav View} and {nav Full Screen}, the thumbnail bar model has already
filtered out any item that isn't viewable, therefore it's easy to simply
get the current index and total items. In {nav Browse}, folders and
archives are displayed, so there needs to be extra logic that filters
them out appropriately.

BUG: 203042
FIXED-IN: 18.08.0

Browse:
{F5811297}

Browse (with selection):
{F5811298}

View:
{F5811299}

View (compare):
{F5811300}

Full Screen:
{F5814068}

Full Screen (large thumbnails):
{F5814069}

Full Screen (thumbnail bar disabled):
{F5814071}

Test Plan:
Browse:
  - Only images and videos should be counted (directories and archives
  ignored).
  - Label should change to "Selected..." when 2 or more images/videos
  selected

View:
  - Count/index should match the thumbnail view, i.e., only items in the
  thumbnail bar should be counted.
  - Changes to documents should be reflected in the index and total:
    - Renaming causing a change in sort order
    - Adding/deleting documents
  - Document count label should be hidden in Compare/Lighttable mode to
  allow room for the Synchronize checkbox

Full Screen View:
  - Same as View above, plus;
  - Enable/disable the thumbnail bar
  - Enable and resize the thumbnail bar
  - View images with long filesnames, and/or addition meta info enabled
  - In Compare/Lighttable mode, current document index should reflect the
  currently selected image

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: rkflx, muhlenpfordt

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D12301