Fix duplicate and markup displaying failed-save message box
Summary:
When saving of a file fails, two dialogs are displayed. One of them
contains markup tags like `<html>` and `<tt>`.
The duplicate message box is caused by KJob's (not disabled) internal
error handling. The markup is displayed because of using strings with
markup as argument for `xi18n*` functions without special handling.

BUG: 393170
FIXED-IN: 18.04.1

Test Plan:
1. Open Gwenview in View or Browse Mode
2. {nav File > Save As} or {key Ctrl Shift S}
3. Try to save in a folder you don't have write permission (e.g. Root)
4. Error message box should display only once and without markup tags

1. Open Gwenview in View or Browse Mode
2. Edit a readonly file (e.g. rotate with {key Ctrl R})
3. Click {nav Save} or {nav Save All}
4. Error message box should display only once and without markup tags

Reviewers: #gwenview, rkflx

Reviewed By: #gwenview, rkflx

Subscribers: rkflx

Differential Revision: https://phabricator.kde.org/D12302